Add Voodoo Sound Support into kernel
Build as module by default as to not interfere with alsa.

Since the kernels have full alsa and wm8903 ioctl registers support
the Voodoo module is not needed to set the analog amp levels.
Alsa and ioctl giving access to everything in the WM8903 chip,
compared to just the analog amp level with the voodoo module.

Change-Id: I3e572b58708007d7a832d9b8c3b2162a11dc4a7b