Set optimization level to 0 by default
Also fix the issue of non-linaro-toolchains not recognizing the -mno-unaligned-access flag, which is required for linaro-compiled kernels to boot up.

Setting it to EOS>0 will cause to build to error out if compiling with a non-linaro-toolchain. So we just set this value to 0, everyone willing to compile using a linaro-gcc can adjust it by him-/herself.

Change-Id: I5464b93cbcff2b462e16765b37a339a3c82f07b4